Restructured a few functional component folders. Error with logos can be fixed with correct paths
