+++ src/app/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import "./App.css";
 import RoutesContainer from "../components/routes/RoutesContainer";
-import { Navbar, Footer } from "../components";
+import { Navbar, Footer } from "../components/functionComponents";
 
 class App extends Component {
   constructor() {
