+++ src/components/functionComponents/Banner.js
