+++ src/components/functionComponents/Navbar.js
@@ -1,7 +1,8 @@
 import React, { Component } from "react";
-import "./views/styles/Navbar.css";
+import "../views/styles/Navbar.css";
+
 import { Link } from "react-router-dom";
-import logo from "../images/Logo.png";
+// import logo from "src/images/Logo.png";
 import { FiAlignCenter } from "react-icons/fi";
 
 class Navbar extends Component {
@@ -19,7 +20,7 @@ class Navbar extends Component {
         <div className="nav-center">
           <div className="nav_header">
             <Link to="/">
-              <img src={logo} alt="logo" class="logo" width="75%" height="75%" />
+              <img src='' alt="logo" class="logo" width="75%" height="75%" />
             </Link>
             <button
               type="button"
