+++ src/components/views/HomePageView.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
 import { render } from "@testing-library/react";
-import { About, Reservation } from "../";
+import { About, Reservation } from "../functionComponents";
 
 const HomePageView = (props) => {
   return (
