+++ src/components/views/MenuPageView.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { Link } from "react-router-dom";
 import { render } from "@testing-library/react";
-import { Banner } from "../";
+import { Banner } from "../functionComponents";
 
 const MenuPageView = (props) => {
   return (
